+++ NEWS.md
@@ -1,4 +1,7 @@
 # evavelo 0.3.0
+
+- Add button in the UI to download full file and logs
+
 - Add update_wb() function
 
 # evavelo 0.2.0